login flow
All of the registration flow happens in a bottom sheet. A user is considered authorized when they have verified their email.
In case a user interrupts the sign up flow after verifying email, they will upon new login have to fill username and password and will be redirected to that page.